refactoring pass to separate logical layer in renderer process
