+++ src/electron/renderer/index.ts
@@ -15,1113 +15,9 @@ if (IS_DEV) {
     cr.consoleRedirect("r2:navigator#electron/renderer/index", process.stdout, process.stderr, true);
 }
 
-import { URL } from "url";
-
-import { Locator, LocatorLocations } from "@r2-shared-js/models/locator";
-import { Publication } from "@r2-shared-js/models/publication";
-import { Link } from "@r2-shared-js/models/publication-link";
-import { encodeURIComponent_RFC3986 } from "@r2-utils-js/_utils/http/UrlUtils";
-import { debounce } from "debounce";
-import * as debug_ from "debug";
-import { ipcRenderer, shell } from "electron";
-
-import { IDocInfo } from "../common/document";
-import {
-    IEventPayload_R2_EVENT_DEBUG_VISUALS,
-    IEventPayload_R2_EVENT_LINK,
-    IEventPayload_R2_EVENT_LOCATOR_VISIBLE,
-    IEventPayload_R2_EVENT_PAGE_TURN,
-    IEventPayload_R2_EVENT_READING_LOCATION,
-    IEventPayload_R2_EVENT_READIUMCSS,
-    IEventPayload_R2_EVENT_SCROLLTO,
-    IEventPayload_R2_EVENT_SHIFT_VIEW_X,
-    IEventPayload_R2_EVENT_TTS_CLICK_ENABLE,
-    IEventPayload_R2_EVENT_TTS_DO_PLAY,
-    R2_EVENT_DEBUG_VISUALS,
-    R2_EVENT_LINK,
-    R2_EVENT_LOCATOR_VISIBLE,
-    R2_EVENT_PAGE_TURN,
-    R2_EVENT_PAGE_TURN_RES,
-    R2_EVENT_READING_LOCATION,
-    R2_EVENT_READIUMCSS,
-    R2_EVENT_SCROLLTO,
-    R2_EVENT_SHIFT_VIEW_X,
-    R2_EVENT_TTS_CLICK_ENABLE,
-    R2_EVENT_TTS_DO_NEXT,
-    R2_EVENT_TTS_DO_PAUSE,
-    R2_EVENT_TTS_DO_PLAY,
-    R2_EVENT_TTS_DO_PREVIOUS,
-    R2_EVENT_TTS_DO_RESUME,
-    R2_EVENT_TTS_DO_STOP,
-    R2_EVENT_TTS_IS_PAUSED,
-    R2_EVENT_TTS_IS_PLAYING,
-    R2_EVENT_TTS_IS_STOPPED,
-} from "../common/events";
-import { IPaginationInfo } from "../common/pagination";
-import { ISelectionInfo } from "../common/selection";
-import {
-    R2_SESSION_WEBVIEW,
-    READIUM2_ELECTRON_HTTP_PROTOCOL,
-    convertCustomSchemeToHttpUrl,
-    convertHttpUrlToCustomScheme,
-} from "../common/sessions";
-import {
-    URL_PARAM_CSS,
-    URL_PARAM_DEBUG_VISUALS,
-    URL_PARAM_EPUBREADINGSYSTEM,
-    URL_PARAM_GOTO,
-    URL_PARAM_PREVIOUS,
-} from "./common/url-params";
-import { highlightsHandleIpcMessage } from "./highlight";
-import { INameVersion } from "./webview/epubReadingSystem";
-import { IReadiumElectronBrowserWindow, IReadiumElectronWebview } from "./webview/state";
-
-import URI = require("urijs");
-
-// import { registerProtocol } from "@r2-navigator-js/electron/renderer/common/protocol";
-// registerProtocol();
-
-const ENABLE_WEBVIEW_RESIZE = true;
-
-// const CLASS_POS_RIGHT = "r2_posRight";
-// const CLASS_SHIFT_LEFT = "r2_shiftedLeft";
-// const CLASS_ANIMATED = "r2_animated";
-
-const ELEMENT_ID_SLIDING_VIEWPORT = "r2_navigator_sliding_viewport";
-
-const debug = debug_("r2:navigator#electron/renderer/index");
-
-// const queryParams = getURLQueryParams();
-
-// // tslint:disable-next-line:no-string-literal
-// const publicationJsonUrl = queryParams["pub"];
-// debug(publicationJsonUrl);
-// const publicationJsonUrl_ = publicationJsonUrl.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL) ?
-//     convertCustomSchemeToHttpUrl(publicationJsonUrl) : publicationJsonUrl;
-// debug(publicationJsonUrl_);
-// const pathBase64 = publicationJsonUrl_.replace(/.*\/pub\/(.*)\/manifest.json/, "$1");
-// debug(pathBase64);
-// const pathDecoded = new Buffer(decodeURIComponent(pathBase64), "base64").toString("utf8");
-// debug(pathDecoded);
-// const pathFileName = pathDecoded.substr(
-//     pathDecoded.replace(/\\/g, "/").lastIndexOf("/") + 1,
-//     pathDecoded.length - 1);
-// debug(pathFileName);
-
-// // tslint:disable-next-line:no-string-literal
-// const lcpHint = queryParams["lcpHint"];
-
-function isRTL(/* link: Link | undefined */): boolean {
-    // if (link && link.Properties) {
-    //     if (link.Properties.Direction === "rtl") {
-    //         return true;
-    //     }
-    //     if (typeof link.Properties.Direction !== "undefined") {
-    //         return false;
-    //     }
-    // }
-    if (_publication &&
-        _publication.Metadata &&
-        _publication.Metadata.Direction) {
-        return _publication.Metadata.Direction.toLowerCase() === "rtl"; //  any other value is LTR
-    }
-    return false;
-}
-
-function isFixedLayout(link: Link | undefined): boolean {
-    if (link && link.Properties) {
-        if (link.Properties.Layout === "fixed") {
-            return true;
-        }
-        if (typeof link.Properties.Layout !== "undefined") {
-            return false;
-        }
-    }
-    if (_publication &&
-        _publication.Metadata &&
-        _publication.Metadata.Rendition) {
-        return _publication.Metadata.Rendition.Layout === "fixed";
-    }
-    return false;
-}
-
-let _epubReadingSystemNameVersion: INameVersion = { name: "Readium2", version: "0.0.0" };
-export function setEpubReadingSystemInfo(nv: INameVersion) {
-    _epubReadingSystemNameVersion = nv;
-}
-
-export function __computeReadiumCssJsonMessage(link: Link | undefined): IEventPayload_R2_EVENT_READIUMCSS {
-
-    if (isFixedLayout(link)) {
-        return { setCSS: undefined, isFixedLayout: true };
-    }
-
-    if (!_computeReadiumCssJsonMessage) {
-        return { setCSS: undefined, isFixedLayout: false };
-    }
-
-    const readiumCssJsonMessage = _computeReadiumCssJsonMessage();
-    return readiumCssJsonMessage;
-}
-let _computeReadiumCssJsonMessage: () => IEventPayload_R2_EVENT_READIUMCSS = () => {
-    return { setCSS: undefined, isFixedLayout: false };
-};
-export function setReadiumCssJsonGetter(func: () => IEventPayload_R2_EVENT_READIUMCSS) {
-    _computeReadiumCssJsonMessage = func;
-}
-
-export interface LocatorExtended {
-    locator: Locator;
-    paginationInfo: IPaginationInfo | undefined;
-    selectionInfo: ISelectionInfo | undefined;
-    selectionIsNew: boolean | undefined;
-    docInfo: IDocInfo | undefined;
-}
-
-let _lastSavedReadingLocation: LocatorExtended | undefined;
-export function getCurrentReadingLocation(): LocatorExtended | undefined {
-    return _lastSavedReadingLocation;
-}
-let _readingLocationSaver: ((locator: LocatorExtended) => void) | undefined;
-const _saveReadingLocation = (docHref: string, locator: IEventPayload_R2_EVENT_READING_LOCATION) => {
-    _lastSavedReadingLocation = {
-        docInfo: locator.docInfo,
-        locator: {
-            href: docHref,
-            locations: {
-                cfi: locator.locations.cfi ?
-                    locator.locations.cfi : undefined,
-                cssSelector: locator.locations.cssSelector ?
-                    locator.locations.cssSelector : undefined,
-                position: (typeof locator.locations.position !== "undefined") ?
-                    locator.locations.position : undefined,
-                progression: (typeof locator.locations.progression !== "undefined") ?
-                    locator.locations.progression : undefined,
-            },
-            text: locator.text,
-            title: locator.title,
-        },
-        paginationInfo: locator.paginationInfo,
-        selectionInfo: locator.selectionInfo,
-        selectionIsNew: locator.selectionIsNew,
-    };
-
-    if (IS_DEV) {
-        // debug(">->->", JSON.stringify(_lastSavedReadingLocation, null, "  "));
-        debug(">->->");
-        debug(_lastSavedReadingLocation);
-    }
-
-    if (_readingLocationSaver) {
-        _readingLocationSaver(_lastSavedReadingLocation);
-    }
-
-    // // tslint:disable-next-line:no-floating-promises
-    // (async () => {
-    //     try {
-    //         const visible = await isLocatorVisible(_lastSavedReadingLocation.locator);
-    //         debug(`isLocatorVisible async: ${visible}`);
-    //     } catch (err) {
-    //         debug(err);
-    //     }
-    // })();
-};
-export function setReadingLocationSaver(func: (locator: LocatorExtended) => void) {
-    _readingLocationSaver = func;
-}
-
-export function readiumCssOnOff() {
-
-    const savedReadingLocation = _lastSavedReadingLocation;
-
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (activeWebView) {
-        const payload1 = __computeReadiumCssJsonMessage(activeWebView.READIUM2.link);
-
-        if (activeWebView.style.transform !== "none") {
-            activeWebView.send("R2_EVENT_HIDE");
-
-            setTimeout(() => {
-                shiftWebview(activeWebView, 0, undefined); // reset
-                activeWebView.send(R2_EVENT_READIUMCSS, payload1);
-            }, 10);
-        } else {
-            activeWebView.send(R2_EVENT_READIUMCSS, payload1);
-        }
-    }
-
-    if (savedReadingLocation) {
-        setTimeout(() => {
-            handleLinkLocator(savedReadingLocation.locator);
-        }, 60);
-    }
-}
-
-export async function isLocatorVisible(locator: Locator): Promise<boolean> {
-    return new Promise<boolean>((resolve, reject) => {
-        const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-
-        if (!activeWebView) {
-            reject("No navigator webview?!");
-            return;
-        }
-        if (!activeWebView.READIUM2.link) {
-            reject("No navigator webview link?!");
-            return;
-        }
-        if (activeWebView.READIUM2.link.Href !== locator.href) {
-            // debug(`isLocatorVisible FALSE: ${activeWebView.READIUM2.link.Href} !== ${locator.href}`);
-            resolve(false);
-            return;
-        }
-
-        // const cb = (_event: any, payload: IEventPayload_R2_EVENT_LOCATOR_VISIBLE) => {
-        //     debug("R2_EVENT_LOCATOR_VISIBLE");
-        //     debug(payload.visible);
-        // };
-        // ipcRenderer.once(R2_EVENT_LOCATOR_VISIBLE, cb);
-
-        const cb = (event: Electron.IpcMessageEvent) => {
-            if (event.channel === R2_EVENT_LOCATOR_VISIBLE) {
-                const webview = event.currentTarget as IReadiumElectronWebview;
-                if (webview !== activeWebView) {
-                    reject("Wrong navigator webview?!");
-                    return;
-                }
-                const payloadPong = event.args[0] as IEventPayload_R2_EVENT_LOCATOR_VISIBLE;
-                // debug(`isLocatorVisible: ${payload_.visible}`);
-                activeWebView.removeEventListener("ipc-message", cb);
-                resolve(payloadPong.visible);
-            }
-        };
-        activeWebView.addEventListener("ipc-message", cb);
-        const payloadPing: IEventPayload_R2_EVENT_LOCATOR_VISIBLE = { location: locator.locations, visible: false };
-        activeWebView.send(R2_EVENT_LOCATOR_VISIBLE, payloadPing);
-    });
-}
-
-let _webview1: IReadiumElectronWebview;
-
-let _publication: Publication | undefined;
-let _publicationJsonUrl: string | undefined;
-
-let _rootHtmlElement: HTMLElement | undefined;
-let _slidingViewport: HTMLElement | undefined;
-
-export function handleLink(href: string, previous: boolean | undefined, useGoto: boolean) {
-
-    const special = href.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://");
-    if (special) {
-        loadLink(href, previous, useGoto);
-    } else {
-        const okay = loadLink(href, previous, useGoto);
-        if (!okay) {
-            debug("EXTERNAL LINK:");
-            debug(href);
-            shell.openExternal(href);
-        }
-    }
-}
-
-export function handleLinkUrl(href: string) {
-    handleLink(href, undefined, false);
-}
-
-export function handleLinkLocator(location: Locator | undefined) {
-
-    // installNavigatorDOM() should have initiated the state
-    if (!_publication || !_publicationJsonUrl) {
-        return;
-    }
-
-    let linkToLoad: Link | undefined;
-    let linkToLoadGoto: LocatorLocations | undefined;
-    if (location && location.href) {
-        if (_publication.Spine && _publication.Spine.length) {
-            linkToLoad = _publication.Spine.find((spineLink) => {
-                return spineLink.Href === location.href;
-            });
-            if (linkToLoad && location.locations) {
-                linkToLoadGoto = location.locations;
-            }
-        }
-        if (!linkToLoad &&
-            _publication.Resources && _publication.Resources.length) {
-            linkToLoad = _publication.Resources.find((resLink) => {
-                return resLink.Href === location.href;
-            });
-            if (linkToLoad && location.locations) {
-                linkToLoadGoto = location.locations;
-            }
-        }
-    }
-    if (!linkToLoad) {
-        if (_publication.Spine && _publication.Spine.length) {
-            const firstLinear = _publication.Spine[0];
-            if (firstLinear) {
-                linkToLoad = firstLinear;
-            }
-        }
-    }
-
-    if (linkToLoad) {
-        const useGoto = typeof linkToLoadGoto !== "undefined"
-            // && typeof linkToLoadGoto.cssSelector !== "undefined"
-            ;
-        const uri = new URL(linkToLoad.Href, _publicationJsonUrl);
-        uri.hash = "";
-        uri.search = "";
-        const urlNoQueryParams = uri.toString(); // _publicationJsonUrl + "/../" + linkToLoad.Href;
-        const hrefToLoad = urlNoQueryParams +
-            ((useGoto) ? ("?" + URL_PARAM_GOTO + "=" +
-                encodeURIComponent_RFC3986(new Buffer(JSON.stringify(linkToLoadGoto, null, "")).toString("base64"))) :
-                "");
-        handleLink(hrefToLoad, undefined, useGoto);
-    }
-}
-
-export function installNavigatorDOM(
-    publication: Publication,
-    publicationJsonUrl: string,
-    rootHtmlElementID: string,
-    preloadScriptPath: string,
-    location: Locator | undefined) {
-
-    _publication = publication;
-    _publicationJsonUrl = publicationJsonUrl;
-
-    if (IS_DEV) {
-        debug("||||||++||||| installNavigatorDOM: ", JSON.stringify(location));
-
-        const debugVisualz = (window.localStorage &&
-            window.localStorage.getItem(URL_PARAM_DEBUG_VISUALS) === "true") ? true : false;
-        debug("debugVisuals GET: ", debugVisualz);
-
-        (window as IReadiumElectronBrowserWindow).READIUM2 = {
-            DEBUG_VISUALS: debugVisualz,
-            getActiveWebView: (): IReadiumElectronWebview => {
-                return _webview1;
-
-                // let activeWebView: IReadiumElectronWebview;
-
-                // const slidingViewport = document.getElementById(ELEMENT_ID_SLIDING_VIEWPORT) as HTMLElement;
-                // if (slidingViewport.classList.contains(CLASS_SHIFT_LEFT)) {
-                //     if (_webview1.classList.contains(CLASS_POS_RIGHT)) {
-                //         activeWebView = _webview1;
-                //     } else {
-                //         activeWebView = _webview2;
-                //     }
-                // } else {
-                //     if (_webview2.classList.contains(CLASS_POS_RIGHT)) {
-                //         activeWebView = _webview1;
-                //     } else {
-                //         activeWebView = _webview2;
-                //     }
-                // }
-
-                // return activeWebView;
-            },
-            publication: _publication,
-            publicationURL: _publicationJsonUrl,
-            ttsClickEnabled: false,
-        };
-
-        (window as any).READIUM2.debug = (debugVisuals: boolean) => {
-            debug("debugVisuals SET: ", debugVisuals);
-            (window as IReadiumElectronBrowserWindow).READIUM2.DEBUG_VISUALS = debugVisuals;
-
-            const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-            if (activeWebView) {
-                const payload: IEventPayload_R2_EVENT_DEBUG_VISUALS
-                    = { debugVisuals };
-                activeWebView.send(R2_EVENT_DEBUG_VISUALS, payload);
-            }
-            if (window.localStorage) {
-                window.localStorage.setItem(URL_PARAM_DEBUG_VISUALS, debugVisuals ? "true" : "false");
-            }
-            setTimeout(() => {
-                // const loc = getCurrentReadingLocation();
-                // debug("|||||||||||||| getCurrentReadingLocation: ", JSON.stringify(loc));
-                if (_lastSavedReadingLocation) {
-                    handleLinkLocator(_lastSavedReadingLocation.locator);
-                }
-            }, 100);
-        };
-
-        (window as any).READIUM2.debugItems =
-            (cssSelector: string, cssClass: string, cssStyles: string | undefined) => {
-
-                if (cssStyles) {
-                    debug("debugVisuals ITEMS: ", `${cssSelector} --- ${cssClass} --- ${cssStyles}`);
-                }
-
-                const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-                // let delay = 0;
-                // if (!(window as IReadiumElectronBrowserWindow).READIUM2.DEBUG_VISUALS) {
-                //     (window as any).READIUM2.debug(true);
-                //     delay = 200;
-                // }
-                // setTimeout(() => {
-                //     if (activeWebView) {
-                //         const payload: IEventPayload_R2_EVENT_DEBUG_VISUALS
-                //             = { debugVisuals: true, cssSelector, cssClass, cssStyles };
-                //         activeWebView.send(R2_EVENT_DEBUG_VISUALS, payload);
-                //     }
-                // }, delay);
-
-                if (activeWebView) {
-                    const d = (window as IReadiumElectronBrowserWindow).READIUM2.DEBUG_VISUALS;
-                    const payload: IEventPayload_R2_EVENT_DEBUG_VISUALS
-                        = { debugVisuals: d, cssSelector, cssClass, cssStyles };
-                    activeWebView.send(R2_EVENT_DEBUG_VISUALS, payload);
-                }
-            };
-    }
-
-    _rootHtmlElement = document.getElementById(rootHtmlElementID) as HTMLElement;
-    if (!_rootHtmlElement) {
-        debug("!rootHtmlElement ???");
-        return;
-    }
-
-    _slidingViewport = document.createElement("div");
-    _slidingViewport.setAttribute("id", ELEMENT_ID_SLIDING_VIEWPORT);
-    _slidingViewport.setAttribute("style", "display: block; position: absolute; left: 0; width: 200%; " +
-        "top: 0; bottom: 0; margin: 0; padding: 0; box-sizing: border-box; background: white; overflow: hidden;");
-
-    _webview1 = createWebView(preloadScriptPath);
-    _webview1.READIUM2 = {
-        id: 1,
-        link: undefined,
-    };
-    _webview1.setAttribute("id", "webview1");
-
-    _slidingViewport.appendChild(_webview1 as Node);
-    // slidingViewport.appendChild(_webview2 as Node);
-
-    _rootHtmlElement.appendChild(_slidingViewport);
-
-    // if (isRTL()) {
-    //     _webview1.classList.add(CLASS_POS_RIGHT);
-    //     _webview1.style.left = "50%";
-    // }
-    // else {
-    //     _webview2.classList.add(CLASS_POS_RIGHT);
-    //     _webview2.style.left = "50%";
-    // }
-
-    setTimeout(() => {
-        handleLinkLocator(location);
-    }, 100);
-}
-
-export function navLeftOrRight(left: boolean) {
-    if (!_publication) {
-        return;
-    }
-    const rtl = isRTL();
-    const goPREVIOUS = left ? !rtl : rtl;
-    const payload: IEventPayload_R2_EVENT_PAGE_TURN = {
-        direction: rtl ? "RTL" : "LTR",
-        go: goPREVIOUS ? "PREVIOUS" : "NEXT",
-    };
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (activeWebView) {
-        activeWebView.send(R2_EVENT_PAGE_TURN, payload); // .getWebContents()
-    }
-}
-
-function loadLink(hrefFull: string, previous: boolean | undefined, useGoto: boolean): boolean {
-
-    if (!_publication || !_publicationJsonUrl) {
-        return false;
-    }
-
-    if (hrefFull.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://")) {
-        hrefFull = convertCustomSchemeToHttpUrl(hrefFull);
-    }
-
-    const pubJsonUri = _publicationJsonUrl.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://") ?
-        convertCustomSchemeToHttpUrl(_publicationJsonUrl) : _publicationJsonUrl;
-
-    let linkPath: string | undefined;
-
-    const urlToLink = new URL(hrefFull);
-    urlToLink.hash = "";
-    urlToLink.search = "";
-    const urlPublication = new URL(pubJsonUri);
-    urlPublication.hash = "";
-    urlPublication.search = "";
-    let iBreak = -1;
-    for (let i = 0; i < urlPublication.pathname.length; i++) {
-        const c1 = urlPublication.pathname[i];
-        if (i < urlToLink.pathname.length) {
-            const c2 = urlToLink.pathname[i];
-            if (c1 !== c2) {
-                iBreak = i;
-                break;
-            }
-        } else {
-            break;
-        }
-    }
-    if (iBreak > 0) {
-        linkPath = urlToLink.pathname.substr(iBreak);
-    }
-
-    if (!linkPath) {
-        return false;
-    }
-
-    // const pubUri = new URI(pubJsonUri);
-    // // "/pub/BASE64_PATH/manifest.json" ==> "/pub/BASE64_PATH/"
-    // const pathPrefix = decodeURIComponent(pubUri.path().replace("manifest.json", ""));
-    // // "/pub/BASE64_PATH/epub/chapter.html" ==> "epub/chapter.html"
-    // const normPath = decodeURIComponent(linkUri.normalizePath().path());
-    // const linkPath = normPath.replace(pathPrefix, "");
-
-    let pubLink = _publication.Spine ? _publication.Spine.find((spineLink) => {
-        return spineLink.Href === linkPath;
-    }) : undefined;
-    if (!pubLink) {
-        pubLink = _publication.Resources.find((spineLink) => {
-            return spineLink.Href === linkPath;
-        });
-    }
-
-    if (!pubLink) {
-        debug("FATAL WEBVIEW READIUM2_LINK ??!! " + hrefFull + " ==> " + linkPath);
-        return false;
-    }
-
-    const linkUri = new URI(hrefFull);
-    linkUri.search((data: any) => {
-        // overrides existing (leaves others intact)
-
-        if (typeof previous === "undefined") {
-            // erase unwanted forward of query param during linking
-            data[URL_PARAM_PREVIOUS] = undefined;
-            // delete data[URL_PARAM_PREVIOUS];
-        } else {
-            data[URL_PARAM_PREVIOUS] = previous ? "true" : "false";
-        }
-
-        if (!useGoto) {
-            // erase unwanted forward of query param during linking
-            data[URL_PARAM_GOTO] = undefined;
-            // delete data[URL_PARAM_GOTO];
-        }
-    });
-    if (useGoto) {
-        linkUri.hash("").normalizeHash();
-    }
-
-    // no need for encodeURIComponent_RFC3986, auto-encoded by URI class
-
-    const rcssJson = __computeReadiumCssJsonMessage(pubLink);
-    const rcssJsonstr = JSON.stringify(rcssJson, null, "");
-    const rcssJsonstrBase64 = new Buffer(rcssJsonstr).toString("base64");
-
-    const rersJson = _epubReadingSystemNameVersion;
-    const rersJsonstr = JSON.stringify(rersJson, null, "");
-    const rersJsonstrBase64 = new Buffer(rersJsonstr).toString("base64");
-
-    linkUri.search((data: any) => {
-        // overrides existing (leaves others intact)
-
-        // tslint:disable-next-line:no-string-literal
-        data[URL_PARAM_CSS] = rcssJsonstrBase64;
-
-        // tslint:disable-next-line:no-string-literal
-        data[URL_PARAM_EPUBREADINGSYSTEM] = rersJsonstrBase64;
-
-        // tslint:disable-next-line:no-string-literal
-        data[URL_PARAM_DEBUG_VISUALS] = (IS_DEV && (window as IReadiumElectronBrowserWindow).READIUM2.DEBUG_VISUALS) ?
-            "true" : "false";
-    });
-
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-
-    if (activeWebView && activeWebView.READIUM2.link === pubLink) {
-        const goto = useGoto ? linkUri.search(true)[URL_PARAM_GOTO] as string : undefined;
-        const hash = useGoto ? undefined : linkUri.fragment(); // without #
-
-        debug("ALREADY LOADED: " + pubLink.Href);
-
-        const payload: IEventPayload_R2_EVENT_SCROLLTO = {
-            goto,
-            hash,
-            previous: previous ? true : false,
-        };
-
-        if (IS_DEV) {
-            const msgStr = JSON.stringify(payload);
-            debug(msgStr);
-        }
-        if (activeWebView) {
-            if (activeWebView.style.transform !== "none") {
-                activeWebView.send("R2_EVENT_HIDE");
-
-                setTimeout(() => {
-                    shiftWebview(activeWebView, 0, undefined); // reset
-                    activeWebView.send(R2_EVENT_SCROLLTO, payload);
-                }, 10);
-            } else {
-                activeWebView.send(R2_EVENT_SCROLLTO, payload);
-            }
-        }
-
-        return true;
-
-        // const webviewToReuse = wv1AlreadyLoaded ? _webview1 : _webview2;
-        // // const otherWebview = webviewToReuse === _webview2 ? _webview1 : _webview2;
-        // if (webviewToReuse !== activeWebView) {
-
-        //     debug("INTO VIEW ...");
-
-        //     const slidingView = document.getElementById(ELEMENT_ID_SLIDING_VIEWPORT) as HTMLElement;
-        //     if (slidingView) {
-        //         let animate = true;
-        //         if (goto || hash) {
-        //             debug("DISABLE ANIM");
-        //             animate = false;
-        //         } else if (previous) {
-        //             if (!slidingView.classList.contains(CLASS_SHIFT_LEFT)) {
-        //                 debug("DISABLE ANIM");
-        //                 animate = false;
-        //             }
-        //         }
-        //         if (animate) {
-        //             if (!slidingView.classList.contains(CLASS_ANIMATED)) {
-        //                 slidingView.classList.add(CLASS_ANIMATED);
-        //                 slidingView.style.transition = "left 500ms ease-in-out";
-        //             }
-        //         } else {
-        //             if (slidingView.classList.contains(CLASS_ANIMATED)) {
-        //                 slidingView.classList.remove(CLASS_ANIMATED);
-        //                 slidingView.style.transition = "none";
-        //             }
-        //         }
-        //         if (slidingView.classList.contains(CLASS_SHIFT_LEFT)) {
-        //             slidingView.classList.remove(CLASS_SHIFT_LEFT);
-        //             slidingView.style.left = "0";
-
-        //             // if (_webview1.classList.contains(CLASS_POS_RIGHT)) {
-        //             //     // activeWebView === _webview1;
-        //             // } else {
-        //             //     // activeWebView === _webview2;
-        //             // }
-        //         } else {
-        //             slidingView.classList.add(CLASS_SHIFT_LEFT);
-        //             slidingView.style.left = "-100%";
-
-        //             // if (_webview2.classList.contains(CLASS_POS_RIGHT)) {
-        //             //     // activeWebView === _webview1;
-        //             // } else {
-        //             //     // activeWebView === _webview2;
-        //             // }
-        //         }
-        //     }
-        // }
-    }
-
-    // if (!isFixedLayout(pubLink)) {
-    //     if (_rootHtmlElement) {
-    //         _rootHtmlElement.dispatchEvent(new Event(DOM_EVENT_HIDE_VIEWPORT));
-    //     }
-    // }
-
-    if (activeWebView) {
-
-        const uriStr = linkUri.toString();
-
-        // if (IS_DEV) {
-        //     debug("####### >>> ---");
-        //     debug(activeWebView.READIUM2.id);
-        //     debug(pubLink.Href);
-        //     debug(uriStr);
-        //     debug(linkUri.hash()); // with #
-        //     debug(linkUri.fragment()); // without #
-        //     // tslint:disable-next-line:no-string-literal
-        //     const gto = linkUri.search(true)[URL_PARAM_GOTO];
-        //     debug(gto ? (new Buffer(gto, "base64").toString("utf8")) : ""); // decodeURIComponent
-        //     // tslint:disable-next-line:no-string-literal
-        //     debug(linkUri.search(true)[URL_PARAM_PREVIOUS]);
-        //     // tslint:disable-next-line:no-string-literal
-        //     debug(linkUri.search(true)[URL_PARAM_CSS]);
-        //     debug("####### >>> ---");
-        // }
-
-        const webviewAlreadyHasContent = (typeof activeWebView.READIUM2.link !== "undefined")
-            && activeWebView.READIUM2.link !== null;
-
-        activeWebView.READIUM2.link = pubLink;
-
-        const needConvert = _publicationJsonUrl.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://");
-        const uriStr_ = uriStr.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://") ?
-            uriStr : (needConvert ? convertHttpUrlToCustomScheme(uriStr) : uriStr);
-        if (IS_DEV) {
-            debug("setAttribute SRC:");
-            debug(uriStr_);
-        }
-
-        if (activeWebView.style.transform !== "none") {
-            // activeWebView.setAttribute("src", "data:, ");
-
-            if (webviewAlreadyHasContent) {
-                activeWebView.send("R2_EVENT_HIDE");
-            }
-
-            setTimeout(() => {
-                shiftWebview(activeWebView, 0, undefined); // reset
-                activeWebView.setAttribute("src", uriStr_);
-            }, 10);
-        } else {
-            activeWebView.setAttribute("src", uriStr_);
-        }
-    }
-
-    // activeWebView.getWebContents().loadURL(uriStr_, { extraHeaders: "pragma: no-cache\n" });
-    // activeWebView.loadURL(uriStr_, { extraHeaders: "pragma: no-cache\n" });
-
-    // ALWAYS FALSE => let's comment for now...
-    // const enableOffScreenRenderPreload = false;
-    // if (enableOffScreenRenderPreload) {
-    //     setTimeout(() => {
-    //         if (!_publication || !pubLink) {
-    //             return;
-    //         }
-
-    //         const otherWebview = activeWebView === _webview2 ? _webview1 : _webview2;
-
-    //         // let inSpine = true;
-    //         const index = _publication.Spine.indexOf(pubLink);
-    //         // if (!index) {
-    //         //     inSpine = false;
-    //         //     index = _publication.Resources.indexOf(pubLink);
-    //         // }
-    //         if (index >= 0 &&
-    //             previous && (index - 1) >= 0 ||
-    //             !previous && (index + 1) < _publication.Spine.length
-    //             // (index + 1) < (inSpine ? _publication.Spine.length : _publication.Resources.length)
-    //         ) {
-    //             const nextPubLink = _publication.Spine[previous ? (index - 1) : (index + 1)];
-    //             // (inSpine ? _publication.Spine[index + 1] : _publication.Resources[index + 1]);
-
-    //             if (otherWebview.READIUM2.link !== nextPubLink) {
-    //                 const linkUriNext = new URI(_publicationJsonUrl + "/../" + nextPubLink.Href);
-    //                 linkUriNext.normalizePath();
-    //                 linkUriNext.search((data: any) => {
-    //                     // overrides existing (leaves others intact)
-
-    //                     // tslint:disable-next-line:no-string-literal
-    //                     data[URL_PARAM_CSS] = rcssJsonstrBase64;
-    //                 });
-    //                 const uriStrNext = linkUriNext.toString();
-
-    //                 debug("####### ======");
-    //                 debug(otherWebview.READIUM2.id);
-    //                 debug(nextPubLink.Href);
-    //                 debug(linkUriNext.hash()); // with #
-    //                 debug(linkUriNext.fragment()); // without #
-    //                 // tslint:disable-next-line:no-string-literal
-    //                 debug(linkUriNext.search(true)[URL_PARAM_GOTO]);
-    //                 // tslint:disable-next-line:no-string-literal
-    //                 debug(linkUriNext.search(true)[URL_PARAM_PREVIOUS]);
-    //                 debug("####### ======");
-    //                 otherWebview.READIUM2.link = nextPubLink;
-    //                 otherWebview.setAttribute("src", uriStrNext);
-    //             }
-    //         }
-    //     }, 300);
-    // }
-
-    return true;
-}
-
-function shiftWebview(webview: IReadiumElectronWebview, offset: number, backgroundColor: string | undefined) {
-    if (!offset) {
-        webview.style.transform = "none";
-        // if (_slidingViewport) {
-        //     _slidingViewport.style.backgroundColor = "white";
-        // }
-    } else {
-        // console.log(`backgroundColor:::::::::: ${backgroundColor}`);
-        if (_slidingViewport && backgroundColor) {
-            _slidingViewport.style.backgroundColor = backgroundColor;
-        }
-        webview.style.transform = `translateX(${offset}px)`;
-    }
-}
-
-function createWebView(preloadScriptPath: string): IReadiumElectronWebview {
-
-    // Unfortunately the Chromium web inspector crashes when closing preload :(
-    // Also, the debugger fails to open the sourcemaps (maybe related issue?)
-    // process.stderr.write("\n####\n" + preloadScriptPath + "\n####\n");
-    // TODO: what are the critical features needed from Node context
-    // that justify using webview.preload? Can we instead use regular DOM code?
-    // The ReadiumCSS injection is now streamer-based (best performance / timing)
-    // and we can use postMessage instead of Electron IPC.
-    // Also, preload really does most of its processing once DOM-ready.
-    // Perhaps the main problem would be exposing the internal logic of navigator
-    // into EPUB content documents? (preload is good for isolating app code)
-
-    const wv = document.createElement("webview");
-    // tslint:disable-next-line:max-line-length
-    // https://github.com/electron/electron/blob/master/docs/tutorial/security.md#3-enable-context-isolation-for-remote-content
-    wv.setAttribute("webpreferences",
-        "nodeIntegration=0, nodeIntegrationInWorker=0, sandbox=0, javascript=1, " +
-        "contextIsolation=0, webSecurity=1, allowRunningInsecureContent=0");
-    wv.setAttribute("partition", R2_SESSION_WEBVIEW);
-    if (_publicationJsonUrl) {
-        // const ref = _publicationJsonUrl.startsWith(READIUM2_ELECTRON_HTTP_PROTOCOL + "://") ?
-        //     _publicationJsonUrl : convertHttpUrlToCustomScheme(_publicationJsonUrl);
-        wv.setAttribute("httpreferrer", _publicationJsonUrl);
-    }
-    wv.setAttribute("style", "display: flex; margin: 0; padding: 0; box-sizing: border-box; " +
-        "position: absolute; left: 0; width: 50%; bottom: 0; top: 0;");
-    wv.setAttribute("preload", preloadScriptPath); // "file://"
-
-    // https://github.com/electron/electron/blob/v3.0.0/docs/api/breaking-changes.md#webview
-    if (ENABLE_WEBVIEW_RESIZE) {
-        wv.setAttribute("disableguestresize", "");
-    }
-
-    setTimeout(() => {
-        wv.removeAttribute("tabindex");
-    }, 500);
-
-    wv.addEventListener("dom-ready", () => {
-        // https://github.com/electron/electron/blob/v3.0.0/docs/api/breaking-changes.md#webcontents
-        // wc.openDevTools({ mode: "detach" });
-        wv.clearHistory();
-
-        if ((window as IReadiumElectronBrowserWindow).READIUM2) {
-            ttsClickEnable((window as IReadiumElectronBrowserWindow).READIUM2.ttsClickEnabled);
-        }
-    });
-
-    wv.addEventListener("ipc-message", (event: Electron.IpcMessageEvent) => {
-        const webview = event.currentTarget as IReadiumElectronWebview;
-        const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-        if (webview !== activeWebView) {
-            return;
-        }
-
-        if (event.channel === R2_EVENT_SHIFT_VIEW_X) {
-            shiftWebview(webview,
-                (event.args[0] as IEventPayload_R2_EVENT_SHIFT_VIEW_X).offset,
-                (event.args[0] as IEventPayload_R2_EVENT_SHIFT_VIEW_X).backgroundColor);
-        } else if (event.channel === R2_EVENT_LINK) {
-            // debug("R2_EVENT_LINK (webview.addEventListener('ipc-message')");
-            const payload = event.args[0] as IEventPayload_R2_EVENT_LINK;
-            handleLinkUrl(payload.url);
-        } else if (event.channel === R2_EVENT_READING_LOCATION) {
-            const payload = event.args[0] as IEventPayload_R2_EVENT_READING_LOCATION;
-            if (webview.READIUM2.link && _saveReadingLocation) {
-                // TODO: position metrics, based on arbitrary number of characters (1034)
-                // https://github.com/readium/architecture/tree/master/positions
-                // https://github.com/readium/architecture/tree/master/locators#about-the-notion-of-position
-                // https://github.com/readium/architecture/blob/master/locators/locator-api.md
-                // if (typeof payload.progression !== "undefined" && _publication && webview.READIUM2.link) {
-                //     const totalPositions = _publication.getTotalPositions(webview.READIUM2.link);
-                //     if (totalPositions) {
-                //         payload.position = totalPositions * payload.progression;
-                //         const totalSpinePositions = _publication.getTotalSpinePositions();
-                //         if (totalSpinePositions) {
-                //             payload.globalProgression = payload.position / totalSpinePositions;
-                //         }
-                //     }
-                // }
-                _saveReadingLocation(webview.READIUM2.link.Href, payload);
-            }
-        } else if (event.channel === R2_EVENT_PAGE_TURN_RES) {
-            if (!_publication) {
-                return;
-            }
-
-            const payload = event.args[0] as IEventPayload_R2_EVENT_PAGE_TURN;
-
-            const goPREVIOUS = payload.go === "PREVIOUS"; // any other value is NEXT
-
-            if (!webview.READIUM2.link) {
-                debug("WEBVIEW READIUM2_LINK ??!!");
-                return;
-            }
-
-            let nextOrPreviousSpineItem: Link | undefined;
-            if (_publication.Spine) {
-                for (let i = 0; i < _publication.Spine.length; i++) {
-                    if (_publication.Spine[i] === webview.READIUM2.link) {
-                        if (goPREVIOUS && (i - 1) >= 0) {
-                            nextOrPreviousSpineItem = _publication.Spine[i - 1];
-                        } else if (!goPREVIOUS && (i + 1) < _publication.Spine.length) {
-                            nextOrPreviousSpineItem = _publication.Spine[i + 1];
-                        }
-                        break;
-                    }
-                }
-            }
-            if (!nextOrPreviousSpineItem) {
-                return;
-            }
-            if (_publicationJsonUrl) {
-                const uri = new URL(nextOrPreviousSpineItem.Href, _publicationJsonUrl);
-                uri.hash = "";
-                uri.search = "";
-                const urlNoQueryParams = uri.toString(); // _publicationJsonUrl + "/../" + nextOrPreviousSpineItem.Href;
-                handleLink(urlNoQueryParams, goPREVIOUS, false);
-            }
-        } else if (event.channel === R2_EVENT_TTS_IS_PAUSED) {
-            if (_ttsListener) {
-                _ttsListener(TTSStateEnum.PAUSED);
-            }
-        } else if (event.channel === R2_EVENT_TTS_IS_STOPPED) {
-            if (_ttsListener) {
-                _ttsListener(TTSStateEnum.STOPPED);
-            }
-        } else if (event.channel === R2_EVENT_TTS_IS_PLAYING) {
-            if (_ttsListener) {
-                _ttsListener(TTSStateEnum.PLAYING);
-            }
-        } else if (event.channel === R2_EVENT_LOCATOR_VISIBLE) {
-            // ignore (handled elsewhere)
-        } else {
-            if (!highlightsHandleIpcMessage(event.channel, event.args, webview)) {
-                debug("webview1 ipc-message");
-                debug(event.channel);
-            }
-        }
-    });
-
-    return wv as IReadiumElectronWebview;
-}
-if (ENABLE_WEBVIEW_RESIZE) {
-    // https://github.com/electron/electron/blob/v3.0.0/docs/api/breaking-changes.md#webcontents
-    // https://github.com/electron/electron/blob/v3.0.0/docs/api/breaking-changes.md#webview
-    // wv.setAttribute("disableguestresize", "");
-    const adjustResize = (webview: IReadiumElectronWebview) => {
-        // https://javascript.info/size-and-scroll
-        // offsetW/H: excludes margin, includes border, scrollbar, padding.
-        // clientW/H: excludes margin, border, scrollbar, includes padding.
-        // scrollW/H: like client, but includes hidden (overflow) areas
-        const width = webview.clientWidth;
-        const height = webview.clientHeight;
-        const wc = webview.getWebContents();
-        if (wc && (wc as any).setSize && width && height) {
-            (wc as any).setSize({ // wc is WebContents, works in Electron < 3.0
-                normal: {
-                    height,
-                    width,
-                },
-            });
-        }
-    };
-    const onResizeDebounced = debounce(() => {
-        const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-        if (activeWebView) {
-            adjustResize(activeWebView);
-        }
-    }, 200);
-    window.addEventListener("resize", () => {
-        // const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-        // if (!isFixedLayout(activeWebView.READIUM2.link)) {
-        //     if (_rootHtmlElement) {
-        //         _rootHtmlElement.dispatchEvent(new Event(DOM_EVENT_HIDE_VIEWPORT));
-        //     }
-        // }
-        onResizeDebounced();
-    });
-}
-
-// see webview.addEventListener("ipc-message", ...)
-// needed for main process browserWindow.webContents.send()
-ipcRenderer.on(R2_EVENT_LINK, (_event: any, payload: IEventPayload_R2_EVENT_LINK) => {
-    debug("R2_EVENT_LINK (ipcRenderer.on)");
-    debug(payload.url);
-    handleLinkUrl(payload.url);
-});
-
-export function ttsPlay() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-
-    let startElementCSSSelector: string | undefined;
-    if (_lastSavedReadingLocation && activeWebView.READIUM2 && activeWebView.READIUM2.link) {
-        if (_lastSavedReadingLocation.locator.href === activeWebView.READIUM2.link.Href) {
-            startElementCSSSelector = _lastSavedReadingLocation.locator.locations.cssSelector;
-        }
-    }
-
-    const payload: IEventPayload_R2_EVENT_TTS_DO_PLAY = {
-        rootElement: "html > body", // window.document.body
-        startElement: startElementCSSSelector,
-    };
-    activeWebView.send(R2_EVENT_TTS_DO_PLAY, payload);
-}
-
-export function ttsPause() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-    activeWebView.send(R2_EVENT_TTS_DO_PAUSE);
-}
-export function ttsStop() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-    activeWebView.send(R2_EVENT_TTS_DO_STOP);
-}
-export function ttsResume() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-    activeWebView.send(R2_EVENT_TTS_DO_RESUME);
-}
-export function ttsPrevious() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-    activeWebView.send(R2_EVENT_TTS_DO_PREVIOUS);
-}
-export function ttsNext() {
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-    activeWebView.send(R2_EVENT_TTS_DO_NEXT);
-}
-
-export enum TTSStateEnum {
-    PAUSED = "PAUSED",
-    PLAYING = "PLAYING",
-    STOPPED = "STOPPED",
-}
-let _ttsListener: ((ttsState: TTSStateEnum) => void) | undefined;
-export function ttsListen(ttsListener: (ttsState: TTSStateEnum) => void) {
-    _ttsListener = ttsListener;
-}
-
-export function ttsClickEnable(doEnable: boolean) {
-
-    if ((window as IReadiumElectronBrowserWindow).READIUM2) {
-        (window as IReadiumElectronBrowserWindow).READIUM2.ttsClickEnabled = doEnable;
-    }
-
-    const activeWebView = (window as IReadiumElectronBrowserWindow).READIUM2.getActiveWebView();
-    if (!activeWebView) {
-        return;
-    }
-
-    const payload: IEventPayload_R2_EVENT_TTS_CLICK_ENABLE = {
-        doEnable,
-    };
-    activeWebView.send(R2_EVENT_TTS_CLICK_ENABLE, payload);
-}
+export * from "./dom";
+export * from "./readium-css";
+export * from "./epubReadingSystem";
+export * from "./location";
+export * from "./tts";
+export * from "./highlight";
